+++ python/paddle/fluid/dygraph/dygraph_to_static/__init__.py
@@ -18,10 +18,6 @@ from .static_analysis import *
 from . import variable_trans_func
 from .variable_trans_func import *
 
-from . import convert_call_func
-from .convert_call_func import *
-
-from . import convert_operators
 
 from . import logging_utils
 from .logging_utils import *
@@ -29,5 +25,4 @@ from .logging_utils import *
 __all__ = []
 __all__ += static_analysis.__all__
 __all__ += variable_trans_func.__all__
-__all__ += convert_call_func.__all__
 __all__ += logging_utils.__all__
