+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_convert_call.py
@@ -20,7 +20,7 @@ import numpy as np
 import paddle
 import paddle.fluid as fluid
 from paddle.jit import ProgramTranslator
-from paddle.fluid.dygraph.dygraph_to_static.convert_call_func import (
+from paddle.jit.dy2static.convert_call_func import (
     CONVERSION_OPTIONS,
 )
 from test_program_translator import get_source_code
